refactor: create general controller for request funcs
