+++ types.ts
@@ -1,8 +1,3 @@
-export interface IUser {
-  email: string;
-  password: string;
-}
-
 export type LoginInputs = {
   email: string;
   password: string;
